Remove /opt/glite as the default location for blahp files. HTCONDOR-397
Instead, use the locations where the linux native packages place the
files. The updated defaults for the directories are as follows:
etc: /etc
bin: /usr/bin
sbin: /usr/sbin
libexec: /usr/libexec/blahp

If BLAHPD_LOCATION or GLITE_LOCATION is set in the environment, they are
still used as before.